new work with db

+++ src/services/mcc.py
@@ -1,10 +1,7 @@
-from services.db import DB
 from services.misc import modernize_date, api_ok, api_fail, inject_db
 
 
-db = DB()
-
-
+@inject_db
 def mcc_dashboard(self, params):
     """ no params"""
     flights = self.db.fetchAll("""
@@ -24,7 +21,7 @@ def mcc_dashboard(self, params):
                 if flight.get('node_name', '') != '{model_name}-{node_id}'.format(**flight)
                 else flight.get('node_name', '')
             )
-        del(flight['node_name'], flight['model_name'])
+        del (flight['node_name'], flight['model_name'])
     crews = self.db.fetchAll("""
     select f.flight_id, f.role, u.name, u.id user_id
 from flight_crews f 
@@ -37,48 +34,53 @@ join users u on f.user_id = u.id
     return flights
 
 
+@inject_db
 def mcc_set_crew(self, params):
     """ params = {"flight_id": 1, "role": "pilot", "user_id": 2} """
-    db.query("""delete from flight_crews 
+    self.db.query("""delete from flight_crews 
         where flight_id = :flight_id and (role=:role or user_id=:user_id)""", params, need_commit=True)
-    db.insert('flight_crews', params)
+    self.db.insert('flight_crews', params)
     return api_ok()
 
 
+@inject_db
 def mcc_add_passenger(self, params):
     """ params = {"flight_id": 1, "user_id": 2} """
-    cnt = db.fetchOne('select count(*) from flight_crews where flight_id=:flight_id and role="_other"', params)
+    cnt = self.db.fetchOne('select count(*) from flight_crews where flight_id=:flight_id and role="_other"', params)
     if cnt >= 5:
         return api_fail("В полет разрешается брать не более 5 пассажиров!")
     params['role'] = "_other"
-    db.query("""delete from flight_crews 
+    self.db.query("""delete from flight_crews 
         where flight_id = :flight_id and user_id=:user_id""", params, need_commit=True)
-    db.insert('flight_crews', params)
+    self.db.insert('flight_crews', params)
     return api_ok()
 
 
+@inject_db
 def mcc_set_all_crew(self, params):
     """ params = {"flight_id": int, "crew": [{"role": "pilot", "user_id": 1}, {"role": "radist", "user_id": 2}] }"""
-    if not db.fetchRow('select * from flights where id=:flight_id', params):
+    if not self.db.fetchRow('select * from flights where id=:flight_id', params):
         return api_fail("Полет с указанным номером не существует")
-    db.query("delete from flight_crews where flight_id=:flight_id", params, need_commit=True)
+    self.db.query("delete from flight_crews where flight_id=:flight_id", params, need_commit=True)
     if params['crew']:
         for item in params['crew']:
             item['flight_id'] = params['flight_id']
-        db.insert('flight_crews', params['crew'])
+        self.db.insert('flight_crews', params['crew'])
     return api_ok(crew=params['crew'])
 
 
+@inject_db
 def mcc_remove(self, params):
     """ params = {"flight_id": 1, "user_id": 2} """
-    deleted = db.query("delete from flight_crews where flight_id = :flight_id and user_id = :user_id",
-                       params, need_commit=True)
-    return api_ok(deleted=db.cursor.rowcount)
+    deleted = self.db.query("delete from flight_crews where flight_id = :flight_id and user_id = :user_id",
+                            params, need_commit=True)
+    return api_ok(deleted=self.db.cursor.rowcount)
 
 
+@inject_db
 def mcc_add_flight(self, params):
-    """ params = {"departure": "2018-08-16 15:00:00", "dock": 2} """
-    params['flight_id'] = db.insert('flights', params)
+    """ params = {"departure": "2018-08-16 15:00:00", "dock": 2, "company": "pre"} """
+    params['flight_id'] = self.db.insert('flights', params)
     return api_ok(flight=params)
 
 
@@ -95,9 +97,11 @@ def get_nearest_flight_for_role(self, params):
     return flight
 
 
+@inject_db
 def get_nearest_flight_for_supercargo(self, user_id):
     return get_nearest_flight_for_role(self, {"user_id": user_id, "role": 'supercargo'})
 
 
+@inject_db
 def get_nearest_flight_for_engineer(self, user_id):
     return get_nearest_flight_for_role(self, {"user_id": user_id, "role": 'engineer'})
